server/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.8
Django==5.0.6
django-anymail==10.3
django-cloudinary-storage==0.3.0
django-filter==24.2
django-jazzmin==3.0.0
django-mdeditor==0.1.20
django-oauth-toolkit==2.4.0
djangorestframework==3.15.1
djangorestframework-csv==3.0.2
drf-excel==2.4.0
drf-redesign==0.3.9
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
et-xmlfile==1.1.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
openpyxl==3.1.3
pillow==10.3.0
pycparser==2.22
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
